test: rearrange the MSW mocks
LIIKUNTA-577.
Map LinkedEvent and CMS queries and responses under their own folders.
Improve the mocks so they would be more reusable.
