+++ src/mocks/queries/cms/page.ts
@@ -1,7 +1,7 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { graphql } from 'msw';
 
-import kukkuuPageDemosivu from '../responses/page/kukkuu-page-demosivu.json';
+import kukkuuPageDemosivu from '../../responses/cms/page/kukkuu-page-demosivu.json';
 
 export const queryPage = () =>
   graphql.query('Page', (req, res, ctx) => {
